feat(blog): add new blog post about typesafe queries for Redis
This commit adds a new blog post titled "Typesafe Queries for Redis" to the `data/blog` directory. The blog post discusses the limitations of querying data in Redis and introduces the `@upstash/query` TypeScript SDK as a solution. It explains the concepts of collections and indexes and provides a quick start guide for using the SDK. The commit also includes a demo video and additional resources for further exploration.

The blog post can be found [here](https://github.com/upstash/examples/tree/main/examples/upstash-query).

A new video file, `query-demo.mp4`, is also added to the `public/blog/upstash-query` directory.